+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 <!-- CHANGELOG NEWEST -->
 - IMPROVE performance of ChangeEventBuffer by processing events in bulks.
 - IMPROVE performance of DocCache by processing events in bulks.
-- IMPROVE perofrmance of ChangeEventBuffer and DocCache by lazily processing tasks.
+- IMPROVE performance of ChangeEventBuffer and DocCache by lazily processing tasks.
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->